wp2/t2.1/v1.0/docker/linux/jenkins/.jen/jobs/diaspora_to_csv/workspace/requirements.txt

mysql_connector
pandas
rdflib==6.1.1
requests==2.27.1
SPARQLWrapper==2.0.0